requirements.txt
flask
tensorflow-cpu
numpy
pillow
scikit-learn
matplotlib
seaborn
gunicorn